// CityBoard/cityboard-front/src/models/adverts.ts
/**
 * Generated by orval v6.11.1 🍺
 * Do not edit manually.
 * CityBoard API
 * OpenAPI spec version: v1
 */
import type { AdvertType } from './advertType';
import type { AdvertStatus } from './advertStatus';

export interface Adverts {
  id?: number;
  type?: AdvertType;
  email?: string;
  phone?: string;
  status?: AdvertStatus;
  city?: string;
  district?: string;
  street?: string;
  houseCode?: string;
  flatNumber?: number;
  floor?: number;
  floors?: number;
  roomsNumber?: number;
  area?: number;
  livingArea?: number;
  price?: number;
  description?: string;
  modCheck?: boolean;
  authorId?: number;
}